Create profile for detecting heroku build.
During the build on heroku, the .git directory is removed from the
repository. For this reason the approach with buildnumber-maven-plugin
does not work, since it relies on the .git directory being present.
Luckly on heroku the SOURCE_VERSION environment variable will be
present and it will hold the commit hash of the commit being build.

See https://devcenter.heroku.com/changelog-items/630 for more information.